split part of app.js into handleAvtions.js
